- pb compil visual 2012 64

+++ src/bitm/font_bitm.cpp
@@ -161,7 +161,7 @@ Im2D_Bits<1> cElBitmFont::MultiLineImageString(const std::string & aStrInit,Pt2d
          aSomY += aSz.y;
          aLB.push_back(anIm);
      }
-     aSomY += aSpace.y*(aLStr.size()-1);
+     aSomY += aSpace.y*(int)(aLStr.size()-1);
      if (aRab.x<0)  // Alors Rab est une taille totale
      {
         aRab = - aRab;
